Made quite a few changes.
First, this now uses exceptions to notify the user of problems instead of printing to the error stream.

Added documentation, yay!

Fixed how descenders are printed, namely their open and close tokens are now correctly handled.

Changed the lexing method to allow for a getNextToken() method and implemented that method.