This is pretty much a complete overhaul of the API structure.  However, it is still fully backwards compatible.
Basically, all of the actual processing stuff was shifted to abstract classes and moved to the lexer.abstractLexer package, except for Type because it doesn’t need to be abstracted or shifted.

However, the lexer package still includes the original classes, the only difference is that they extend the abstract ones, hence, the backwards compatibility.